Class 4
In this class, I maked Home Screen, where has the form for subscribe in platform, and maked the connection with graphCMS for save the informations about form.
I created a file, ".env.local" where have keys for access graphCMS API, is important to do that, because the token is an information sensitive.